+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Chg #160: Change PHP constraint in `composer.json` from `php` to `php-64bit`.
 - New #155: Add `NativeTransport` that used native `file_get_contents()` and `file_put_contents()` functions.
 - New #162: Add `ChecklistTask`, `Checklist`, `InputChecklistTask`, `InputChecklist` and `ChecklistTasksDone` types.
-- New #162: Add `checklist` field to `Message` and `ExternalReplyInfo` types.
+- New #162: Add `checklist` and `checklistTasksDone` fields to `Message` type.
+- New #162: Add `checklist` field to `ExternalReplyInfo` type.
 
 ## 0.9.0 April 12, 2025